Create SocialAnalyticsTestHelpers
