+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Analytics/SocialAnalytics/TestHelpers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6075c3d396853d44282902a807874bbe
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
