+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Analytics/SocialAnalytics/TestHelpers/SocialAnalyticsTestHelpers.asmdef
@@ -1,0 +1,24 @@
+{
+    "name": "SocialAnalyticsTestHelpers",
+    "rootNamespace": "",
+    "references": [
+        "UnityEngine.TestRunner",
+        "UnityEditor.TestRunner",
+        "SocialAnalytics"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": true,
+    "precompiledReferences": [
+        "nunit.framework.dll",
+        "NSubstitute.dll",
+        "System.Threading.Tasks.Extensions.dll"
+    ],
+    "autoReferenced": false,
+    "defineConstraints": [
+        "UNITY_INCLUDE_TESTS"
+    ],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
