+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Analytics/SocialAnalytics/TestHelpers/SocialAnalyticsTestHelpers.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c0baddd890750a44b82034a198d1c4af
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
