+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Analytics/SocialAnalytics/TestHelpers/SocialAnalyticsTestHelpers.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5806202a30b8d3845913aa1f3596f0db
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
