+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/FriendsHUD/Tests/FriendsHUDTests.asmdef
@@ -24,7 +24,8 @@
         "GUID:029e05fd02d6c49d5ac9dd7e3725a35b",
         "GUID:d7f1765642f0c22cda30addd0275b5e1",
         "GUID:1e6b57fe78f7b724e9567f29f6a40c2c",
-        "GUID:55b66aa56b81d2d4cafb4a5f02bc90ca"
+        "GUID:55b66aa56b81d2d4cafb4a5f02bc90ca",
+        "GUID:c0baddd890750a44b82034a198d1c4af"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
