+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/PlayerInfoCardHUD/Tests/PlayerInfoCardHUDTests.asmdef
@@ -23,7 +23,8 @@
         "GUID:dffbb581f2650ea4990781f603ac258a",
         "GUID:0663fad624d836944b40ae27c3414652",
         "GUID:3b80b0b562b1cbc489513f09fc1b8f69",
-        "GUID:55b66aa56b81d2d4cafb4a5f02bc90ca"
+        "GUID:55b66aa56b81d2d4cafb4a5f02bc90ca",
+        "GUID:c0baddd890750a44b82034a198d1c4af"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
