+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/PrivateChatWindow/Tests/PrivateChatWindowHUDShould.cs
@@ -1,12 +1,12 @@
-using System;
+using DCL.Helpers;
 using DCL.Interface;
+using NSubstitute;
 using NUnit.Framework;
+using SocialFeaturesAnalytics.TestHelpers;
+using System;
 using System.Collections;
 using System.Collections.Generic;
-using DCL.Helpers;
-using NSubstitute;
 using UnityEngine.TestTools;
-using SocialFeaturesAnalytics;
 
 public class PrivateChatWindowHUDShould : IntegrationTestSuite_Legacy
 {
@@ -186,7 +186,7 @@ public class PrivateChatWindowHUDShould : IntegrationTestSuite_Legacy
         notificationsController.Initialize(new NotificationHUDController());
 
         FriendsHUDController friendsHudController = new FriendsHUDController();
-        friendsHudController.Initialize(new FriendsController_Mock(), SocialAnalytics.CreateMockedSocialAnalytics(), UserProfile.GetOwnUserProfile());
+        friendsHudController.Initialize(new FriendsController_Mock(), SocialAnalyticsTestHelpers.CreateMockedSocialAnalytics(), UserProfile.GetOwnUserProfile());
 
         Assert.IsTrue(view != null, "Friends hud view is null?");
         Assert.IsTrue(controller != null, "Friends hud controller is null?");
