+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/PrivateChatWindow/Tests/PrivateChatWindowHUDTests.asmdef
@@ -24,7 +24,8 @@
         "GUID:0acc523941302664db1f4e527237feb3",
         "GUID:029e05fd02d6c49d5ac9dd7e3725a35b",
         "GUID:d7f1765642f0c22cda30addd0275b5e1",
-        "GUID:55b66aa56b81d2d4cafb4a5f02bc90ca"
+        "GUID:55b66aa56b81d2d4cafb4a5f02bc90ca",
+        "GUID:c0baddd890750a44b82034a198d1c4af"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
