+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/TaskbarHUD/Tests/TaskbarHUDTests.asmdef
@@ -27,7 +27,8 @@
         "GUID:d7f1765642f0c22cda30addd0275b5e1",
         "GUID:bd51ff2b035ce4ab4be7c38c43c88889",
         "GUID:1e6b57fe78f7b724e9567f29f6a40c2c",
-        "GUID:55b66aa56b81d2d4cafb4a5f02bc90ca"
+        "GUID:55b66aa56b81d2d4cafb4a5f02bc90ca",
+        "GUID:c0baddd890750a44b82034a198d1c4af"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
